Refactors extension for cross-browser support
This commit restructures the extension to support Chrome, Firefox, and Opera.
It introduces a build script to generate browser-specific manifests and source code bundles.
Additionally, it centralizes shared resources and logic in a 'base' directory,
and moves browser-specific code into respective 'chromium' and 'firefox' directories.
Removes the generated indexed rulesets.